+++ application/dash_application/known_pages_router.py
@@ -1,14 +1,18 @@
 from collections import namedtuple
+import json
+import logging
+import os
 import random
 import dash_bootstrap_components as dbc
 import dash_html_components as html
 import dash_core_components as core
 from dash.dependencies import Input, Output
 
-from .dash_id import init_ids
 from . import pages
 from ..routes import version
 
+
+logger = logging.getLogger(__name__)
 
 """
 This file acts as a router which serves all the pages the Dash app knows about.
@@ -63,10 +67,58 @@ def navbar(current):
     )
 
 
+def load_user_messages(json_path=os.getenv("DISPLAY_USER_MESSAGE")):
+    """
+    Loads the user messages. Raises an exception in the following cases:
+    * Cannot parse JSON file
+    * If message is not a string
+    * If the view of the message does not exist
+    * If multiple messages exist for a single view
+    """
+    if json_path is None:
+        return {}
+
+    if not os.path.isfile(json_path):
+        logger.warning("User message JSON file does not exist")
+        return {}
+
+    def json_parse(pairs):
+        parsed = {}
+
+        for k, v in pairs:
+            if not isinstance(v, str):
+                raise TypeError("User messages must be a string: {}".format(v))
+
+            if k not in pages_info:
+                raise KeyError("Unknown page name: {}".format(k))
+
+            if k in parsed:
+                raise KeyError("Multiple messages for the same page: {}".format(k))
+            else:
+                parsed[k] = v
+
+        return parsed
+
+    with open(json_path, "r") as f:
+        return json.load(f, object_pairs_hook=json_parse)
+
+
+# Messages to the displayed to users in specific views
+# Key is the page name and the value is the message to be displayed
+# If page name does not exist or value is `None` message won't be displayed
+user_message = load_user_messages()
+
+
 # Default layout element (wraps the page layout elements which are returned by the router)
 layout = html.Div([
     core.Location(id='url', refresh=False),
     navbar(default_title),
+    dbc.Alert(
+        id="user_message",
+        is_open=False,
+        color="danger",
+        style={"margin-left": "15px", "margin-right": "15px"},
+    ),
     core.Loading(id='page-content', type='dot'),
     html.Footer(id='footer', children=[
         html.Hr(), 
@@ -113,3 +165,25 @@ def init_callbacks(dash_app):
             page = pages_info[requested]
             return [page.layout(qs), page.dataversion()]
         return '404', None
+
+    @dash_app.callback(
+        [
+            Output('user_message', 'children'),
+            Output('user_message', 'is_open'),
+        ],
+        [
+            Input('url', 'pathname'),
+        ])
+    def display_user_message(path):
+        """
+        If there is a user message associated with the loaded page, display it.
+        """
+        if path is None:
+            return "", False
+
+        requested = path[1:] # drop the leading slash
+        message = user_message.get(requested, None)
+        if message is None:
+            return "", False
+        else:
+            return message, True